Adds secure clips and cute tooltip for files

+++ ClipboardViewer/CopyingToClipboardInterceptor.cs
@@ -4,39 +4,48 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using Logger;
 
 namespace ClipboardViewer
 {
     class CopyingToClipboardInterceptor
     {
         private readonly IClipboardBuferService _clipboardBuferService;
-        private readonly Form _form;
+		private readonly IEqualityComparer<IDataObject> _comparer;
+		private readonly Form _form;
         private readonly IRenderingHandler _renderingHandler;
 
-        public CopyingToClipboardInterceptor(IClipboardBuferService clipboardBuferService, Form form, IRenderingHandler renderingHandler)
+        public CopyingToClipboardInterceptor(IClipboardBuferService clipboardBuferService, Form form, IRenderingHandler renderingHandler, IEqualityComparer<IDataObject> comparer)
         {
             this._clipboardBuferService = clipboardBuferService;
             this._form = form;
             this._renderingHandler = renderingHandler;
+			this._comparer = comparer;
         }
 
         public void DoOnCtrlC()
         {
+			Logger.Logger.Current.Write("On Ctrl+C");
+
             var currentObject = Clipboard.GetDataObject();
 
-            if (!this._clipboardBuferService.IsLastClip(currentObject))
+            if (currentObject.GetFormats().Any() && !this._clipboardBuferService.IsLastClip(currentObject))
             {
-                if (_clipboardBuferService.Contains(currentObject))
+				Logger.Logger.Current.Write("Is not last clip");
+				if (this._clipboardBuferService.Contains(currentObject))
                 {
-                    _clipboardBuferService.RemoveClip(currentObject);
+					Logger.Logger.Current.Write("Already contains this clip");
+					_clipboardBuferService.RemoveClip(currentObject);					
                 }
-
+				
                 if (_clipboardBuferService.GetClips().Count() == 30)
                 {
-                    _clipboardBuferService.RemoveClip(_clipboardBuferService.FirstClip);
+					Logger.Logger.Current.Write("More than maximum count. Need to remove the first clip");
+					_clipboardBuferService.RemoveClip(_clipboardBuferService.FirstClip);
                 }
 
-                _clipboardBuferService.AddClip(currentObject);
+				Logger.Logger.Current.Write("Add Clip");
+				_clipboardBuferService.AddClip(currentObject);
 
                 if (this._form.WindowState != FormWindowState.Minimized && this._form.Visible)
                 {
